:recycle: Move heatmap scripts from C-PAC to CPAC_regtest_pack
