+++ configs/subjects.py
